fix for two issues:
A) When a video is successfully uploaded via the video upload component, it should change the state of the video catalog component such that the current video is the video that is uploaded, and the state of the video player component such that the video being played is the corresponding video from the b2 bucket link.

B) When a video is cleared via the video catalog component, the video player component should also clear.

Fixed these by lifting state from the individual components and having the main page handle the current video