Consolidated CalculationCard with underlying InteractiveCard implementation into one component file.
Reason: Realized the InteractiveCard component isn't used beyond creating a baseline for CalculationCard, and that CalculationCard is an extra component declared/defined/implemented in page.tsx, which violates separation of concerns (page.tsx should be solely focused on rendering elements on page, components should be held separately). Code is cleaner as well; had to create aliases on the original Card element to create InteractiveCard, but can use it directly in ComponentCard implementation.